+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## 1.0.1 - 2018-03-30
+### Added
+- Support for local `credentials.json` file in addition to CLI
+- Support for updating auth0 client files, with diffing
+- Support logging to syslog, and verbose/debug options
+
+### Added
+- Support for loading credentials from `credentials.json` in addition to CLI, for all scripts
+
 ## 1.0.0 - 2018-03-08
 ### Added
 - This ChangeLog